Preliminary integration of chunk (doesn't compile yet)
