+++ src/chunk_view.rs
@@ -1,4 +1,8 @@
-use crate::{neighbors, raw_neighbors, WORLD_WIDTH};
+use crate::element::{FIXED, FLUID};
+use crate::tile::Tile;
+use crate::world::PairwiseMutate;
+use crate::{above, adjacent_x, neighbors, raw_neighbors, WORLD_WIDTH};
+use rand::Rng;
 use std::ops::{Index, IndexMut};
 
 #[derive(Clone, Copy)]
@@ -44,5 +48,75 @@ impl<'a, T> Index<ChunkIndex> for ChunkView<'a, T> {
 impl<'a, T> IndexMut<ChunkIndex> for ChunkView<'a, T> {
     fn index_mut(&mut self, index: ChunkIndex) -> &mut Self::Output {
         &mut self.chunk[index.0]
+    }
+}
+
+pub struct Chunk<'a> {
+    slice: &'a mut [Option<Tile>],
+}
+
+impl<'a> Chunk<'a> {
+    pub fn new(slice: &mut [Option<Tile>]) -> Chunk {
+        Chunk { slice }
+    }
+    pub fn swap(&mut self, i: usize, j: usize) {
+        self.slice.swap(i, j);
+        if let Some(above_source) = above(i) {
+            if let Some(tile) = &mut self.slice[above_source] {
+                tile.paused = false;
+            }
+        }
+    }
+
+    pub fn move_particle(&mut self, source: usize, destination: usize) {
+        let (source_tile, dest_tile) = self.slice.mutate_pair(source, destination);
+        match (source_tile, dest_tile) {
+            //match (world[source].as_mut(), world[destination].as_mut()) {
+            (None, _) => {
+                //Source particle has moved for some other reason - nothing to do
+            }
+            (Some(_), None) => {
+                self.swap(source, destination);
+            }
+            (Some(ref mut s), Some(ref mut d)) => {
+                if adjacent_x(source, destination) {
+                    if d.has_flag(FIXED) {
+                        s.reflect_velocity_x();
+                    } else {
+                        s.elastic_collide_x(d);
+                    }
+                } else
+                /*if adjacent_y(source, destination)*/
+                {
+                    if d.has_flag(FIXED) {
+                        s.reflect_velocity_y();
+                    } else {
+                        s.elastic_collide_y(d);
+                    }
+                }
+                d.paused = false;
+                if d.has_flag(FLUID) && rand::thread_rng().gen_range(0, 2) == 0 {
+                    // Fluids don't collide, they just push through
+                    self.swap(source, destination);
+                }
+                // TODO: Reimplement collision reactions
+                // self.trigger_collision_reactions(source, destination);
+                // self.trigger_collision_side_effects(source, destination);
+            }
+        }
+    }
+}
+
+impl<'a> Index<ChunkIndex> for Chunk<'a> {
+    type Output = Option<Tile>;
+
+    fn index(&self, index: ChunkIndex) -> &Self::Output {
+        &self.slice[index.0]
+    }
+}
+
+impl<'a> IndexMut<ChunkIndex> for Chunk<'a> {
+    fn index_mut(&mut self, index: ChunkIndex) -> &mut Self::Output {
+        &mut self.slice[index.0]
     }
 }
