+++ src/world.rs
@@ -1,4 +1,4 @@
-use crate::chunk_view::ChunkView;
+use crate::chunk_view::{Chunk, ChunkIndex, ChunkView};
 use crate::element::{Element, FIXED, FLUID, GRAVITY, PAUSE_EXEMPT};
 use crate::tile::{ElementState, Tile};
 use crate::{
@@ -59,7 +59,7 @@ fn mutate_neighborhood<T>(slice: &mut [T], index: usize) -> (&mut T, Neighborhoo
     (&mut center[0], Neighborhood::new(before, after))
 }
 
-trait PairwiseMutate {
+pub trait PairwiseMutate {
     type T;
     fn mutate_pair(&mut self, first: usize, second: usize) -> (&mut Self::T, &mut Self::T);
 }
@@ -211,19 +211,19 @@ impl World {
     /// The function receives a slice of the grid and index into the slice
     /// The slice is guaranteed to contain enough tiles to access the given index's
     /// Moore neighborhood
-    pub fn chunked_for_each(&mut self, f: impl Fn(&mut [Option<Tile>], usize) + Sync + Send) {
+    pub fn chunked_for_each(&mut self, f: impl Fn(Chunk, usize) + Sync + Send) {
         self.grid
             .par_chunks_exact_mut(CHUNK_SIZE)
-            .for_each(|chunk| {
+            .for_each(|slice| {
                 for i in CHUNK_MUTATE_START..CHUNK_MUTATE_END {
-                    f(chunk, i);
+                    f(Chunk::new(slice), i);
                 }
             });
 
         let offset_grid = &mut self.grid[CHUNK_MUTATE_SIZE..];
-        offset_grid.par_chunks_mut(CHUNK_SIZE).for_each(|chunk| {
+        offset_grid.par_chunks_mut(CHUNK_SIZE).for_each(|slice| {
             for i in CHUNK_MUTATE_START..CHUNK_MUTATE_END {
-                f(chunk, i);
+                f(Chunk::new(slice), i);
             }
         });
     }
@@ -379,58 +379,6 @@ impl World {
     }
 }
 
-pub struct Chunk<'a> {
-    grid: &'a mut [Option<Tile>],
-}
-
-impl<'a> Chunk<'a> {
-    pub fn swap(&mut self, i: usize, j: usize) {
-        self.grid.swap(i, j);
-        if let Some(above_source) = above(i) {
-            if let Some(tile) = &mut self.grid[above_source] {
-                tile.paused = false;
-            }
-        }
-    }
-
-    pub fn move_particle(&mut self, source: usize, destination: usize) {
-        let (source_tile, dest_tile) = self.grid.mutate_pair(source, destination);
-        match (source_tile, dest_tile) {
-            //match (world[source].as_mut(), world[destination].as_mut()) {
-            (None, _) => {
-                //Source particle has moved for some other reason - nothing to do
-            }
-            (Some(_), None) => {
-                self.swap(source, destination);
-            }
-            (Some(ref mut s), Some(ref mut d)) => {
-                if adjacent_x(source, destination) {
-                    if d.has_flag(FIXED) {
-                        s.reflect_velocity_x();
-                    } else {
-                        s.elastic_collide_x(d);
-                    }
-                } else
-                /*if adjacent_y(source, destination)*/
-                {
-                    if d.has_flag(FIXED) {
-                        s.reflect_velocity_y();
-                    } else {
-                        s.elastic_collide_y(d);
-                    }
-                }
-                d.paused = false;
-                if d.has_flag(FLUID) && rand::thread_rng().gen_range(0, 2) == 0 {
-                    // Fluids don't collide, they just push through
-                    self.swap(source, destination);
-                }
-                // TODO: Reimplement collision reactions
-                // self.trigger_collision_reactions(source, destination);
-                // self.trigger_collision_side_effects(source, destination);
-            }
-        }
-    }
-}
 // TODO: Upgrade chunk into an actual struct, make these chunk_* functions into methods
 pub fn chunk_swap(chunk: &mut [Option<Tile>], i: usize, j: usize) {
     chunk.swap(i, j);
